Add soft mute feature: surpress continuous message notifications
Similar to mute notifications for any chat, this feature, when
activated, can limit message notifications for selected chat to no more
than one in specified time period. So, if a chat is set as soft mute
within 5 minutes and user isn't focusing on this chat, after the first
notification of this chat, user won't be getting more notifications from
that chat (configurable as send without sound or completely hidden),
until 5 minutes have passed or user have focus on the chat, and then it
starts the new cycle.